Move MempoolContext to separate file
