+++ commander/syncer/txs_context.go
@@ -8,19 +8,6 @@ import (
 	st "github.com/Worldcoin/hubble-commander/storage"
 )
 
-type MempoolContext struct {
-	Mempool      *mempool.TxMempool
-	txController *mempool.TxController
-}
-
-func (c *MempoolContext) Commit() {
-	c.txController.Commit()
-}
-
-func (c *MempoolContext) Rollback() {
-	c.txController.Rollback()
-}
-
 func NewMempoolContext(pool *mempool.Mempool) *MempoolContext {
 	txController, txMempool := pool.BeginTransaction()
 	return &MempoolContext{
